Version 3 - myScript_dev2.js
Stopped the timer as soon as all of the pairs are found.  This is done with a variable "timerStopped" which is updated (to true) when the game ends.  The timer only ticks up when timerStopped === false. 

Modified the "Stars" score to only go from 3 to 2 to 1, getting rid of the possibility of a star score of zero.